Use Float32Array throughout (#10359)
* Use Float32Array throughout

* Update bokehjs/test/unit/core/util/ndarray.ts

* Update bokehjs/test/unit/core/util/ndarray.ts

* Update bokehjs/test/unit/core/util/eq.ts

* revert more copy pasta

* fix bad interpolator index type

* restore edit tool test

* use NumberArray consistently

* update baselines with trivial diffs

* update all baselines

+++ bokehjs/src/lib/core/types.ts
@@ -4,8 +4,8 @@ export type Color = string
 
 export {TypedArray} from "./util/ndarray"
 
-export type NumberArray = Float64Array
-export const NumberArray = Float64Array
+export type NumberArray = Float32Array
+export const NumberArray = Float32Array
 
 export type ColorArray = Uint32Array
 export const ColorArray = Uint32Array
