+++ bokehjs/src/lib/core/util/ndarray.ts
@@ -169,14 +169,14 @@ export type NDArrayTypes = {
   "float64": {typed: Float64Array, ndarray: Float64NDArray}
 }
 
-export function ndarray<K extends DataType = "float64">(array: ArrayBuffer | number[], options?: {dtype?: K, shape?: number[]}): NDArrayTypes[K]["ndarray"]
+export function ndarray<K extends DataType = "float32">(array: ArrayBuffer | number[], options?: {dtype?: K, shape?: number[]}): NDArrayTypes[K]["ndarray"]
 export function ndarray<K extends DataType>(array: NDArrayTypes[K]["typed"], options?: {dtype?: K, shape?: number[]}): NDArrayTypes[K]["ndarray"]
 
 export function ndarray(array: ArrayBuffer | TypedArray | number[], options: {dtype?: DataType, shape?: number[]} = {}): NDArray {
   let {dtype} = options
   if (dtype == null) {
     if (array instanceof ArrayBuffer || isArray(array)) {
-      dtype = "float64"
+      dtype = "float32"
     } else {
       dtype = (() => {
         switch (true) {
